backend/requirements.txt
# Core framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
pydantic-settings==2.7.1

# AI / ML
google-generativeai==0.8.4
openai-whisper==20240930
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1

# PDF processing
PyMuPDF==1.25.3

# Media processing
ffmpeg-python==0.2.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Cache
redis==5.2.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
httpx==0.28.1
aiosqlite==0.20.0

# Utilities
aiofiles==24.1.0
python-dotenv==1.0.1
